refactored directory structure.
