+++ lib/widgets/info_card.dart
@@ -1,0 +1,30 @@
+import 'package:flutter/material.dart';
+import 'package:google_fonts/google_fonts.dart';
+
+class InfoCard extends StatelessWidget {
+  final String label;
+  final String description;
+
+  const InfoCard({super.key, required this.label, required this.description});
+
+  @override
+  Widget build(BuildContext context) {
+    final scale = MediaQuery.of(context).textScaleFactor;
+
+    return Container(
+      padding: const EdgeInsets.all(16),
+      decoration: BoxDecoration(color: Colors.grey.shade900, borderRadius: BorderRadius.circular(12)),
+      child: Column(
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          Text(
+            label,
+            style: GoogleFonts.montserrat(fontSize: 24 * scale, fontWeight: FontWeight.bold, color: Colors.orange),
+          ),
+          const SizedBox(height: 8),
+          Text(description, style: GoogleFonts.montserrat(fontSize: 14 * scale, color: Colors.grey.shade300)),
+        ],
+      ),
+    );
+  }
+}
